requirements.txt
# Core ML libraries
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Gradient Boosting
xgboost>=2.0.0
lightgbm>=4.0.0
tabpfn>=0.1.0

# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Transformers for tabular data
transformers>=4.30.0
pytorch-tabnet>=4.0

# Explainability
shap>=0.42.0
lime>=0.2.0
shapiq>=1.0.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0

# Jupyter
ipykernel>=6.23.0

# Additional utilities
tqdm>=4.65.0
joblib>=1.3.0
pyyaml>=6.0
hydra-core>=1.3.0
